+++ common/src/main/res/layout/item_address_chooser.xml
