+++ feature-wallet-impl/src/main/java/jp/co/soramitsu/feature_wallet_impl/presentation/balance/list/changeAccount/AccountChooserBottomSheetDialog.kt
@@ -30,7 +30,7 @@ class AccountChooserBottomSheetDialog(
     }
 
     override fun holderCreator(): HolderCreator<AddressModel> = { parent ->
-        AddressModelHolder(parent.inflateChild(R.layout.item_account_chooser))
+        AddressModelHolder(parent.inflateChild(R.layout.item_address_chooser))
     }
 }
